feat(probe): replace raw perf-event attach with libbpfgo.AttachUprobeWithOpts
Switch the userspace BPF (bpftime) attach path from hand-rolled
perf_event_open + BPF_LINK_CREATE syscalls to
bpf_program__attach_uprobe_opts via the maxgio92 fork of libbpfgo
(PR aquasecurity/libbpfgo#523).

The raw syscall code in bpftime_attach.go is deleted entirely.
attachSingleUprobes now calls p.bpfProg.AttachUprobeWithOpts for each
(offset, cookie) pair and stores the returned BPFLink in p.links,
so cleanup goes through the same Destroy() path as the kernel path.
The rawFds []int field and its unix.Close loop are removed.

go.mod gains a replace directive pointing github.com/aquasecurity/libbpfgo
at maxgio92/libbpfgo@cc9e1ae (v0.0.0-20260324172613-cc9e1ae113eb).
go.sum needs refreshing locally with: go mod tidy

The bpftime link_handler.hpp bpf_cookie patch is still required since
the cookie propagation bug lives in bpftime, not in the Go attach path.
