+++ pkg/probe/probe.go
@@ -3,12 +3,12 @@ package probe
 import (
 	"context"
 	"embed"
+	"fmt"
 	"path/filepath"
 
 	bpf "github.com/aquasecurity/libbpfgo"
 	"github.com/pkg/errors"
 	log "github.com/rs/zerolog"
-	"golang.org/x/sys/unix"
 )
 
 //go:embed output/*
@@ -30,10 +30,6 @@ type Probe struct {
 	bpfMod  *bpf.Module
 	bpfProg *bpf.BPFProg
 	links   []*bpf.BPFLink
-
-	// rawFds holds the perf event and link FDs created by attachUprobeWithCookie
-	// (userspace BPF mode). They must be kept open for the lifetime of the probe.
-	rawFds []int
 
 	EvtBuf *bpf.RingBuffer
 
@@ -176,10 +172,8 @@ func (p *Probe) CloseEventBuf() {
 
 // CloseBPFMod destroys all BPF links (detaching uprobes) and then closes
 // the BPF module. Links must be explicitly destroyed because AttachUprobeMulti
-// returns a BPFLink that is the sole owner of the uprobe attachment - closing
-// the module alone does not detach the probes.
-// In userspace BPF mode, the raw perf event and link FDs opened by
-// attachUprobeWithCookie are closed here instead.
+// and AttachUprobeWithOpts both return a BPFLink that is the sole owner of the
+// uprobe attachment - closing the module alone does not detach the probes.
 // Must be called after CloseEventBuf so the ring buffer poll goroutine has
 // already stopped.
 func (p *Probe) CloseBPFMod() {
@@ -187,11 +181,45 @@ func (p *Probe) CloseBPFMod() {
 		link.Destroy()
 	}
 	p.links = nil
-	for _, fd := range p.rawFds {
-		unix.Close(fd)
-	}
-	p.rawFds = nil
 	if p.bpfMod != nil {
 		p.bpfMod.Close()
 	}
+}
+
+// attachSingleUprobes attaches the probe to each (offset, cookie) pair using
+// bpf_program__attach_uprobe_opts via libbpfgo. This is the path used in
+// userspace BPF mode (bpftime), which supports single uprobes via perf_event_open
+// but silently no-ops BPF_TRACE_UPROBE_MULTI.
+func (p *Probe) attachSingleUprobes(exePath string, offsets, cookies []uint64) error {
+	var firstErr error
+	attached := 0
+
+	for i, offset := range offsets {
+		cookie := cookies[i]
+		link, err := p.bpfProg.AttachUprobeWithOpts(-1, exePath, offset, cookie)
+		if err != nil {
+			p.logger.Debug().
+				Err(err).
+				Uint64("offset", offset).
+				Uint64("cookie", cookie).
+				Msg("failed to attach uprobe with opts")
+			if firstErr == nil {
+				firstErr = err
+			}
+			continue
+		}
+		p.links = append(p.links, link)
+		attached++
+	}
+
+	if attached == 0 && len(offsets) > 0 {
+		return fmt.Errorf("all %d uprobe attachments failed (first error: %w)", len(offsets), firstErr)
+	}
+
+	p.logger.Debug().
+		Int("attached", attached).
+		Int("total", len(offsets)).
+		Msg("single uprobes attached")
+
+	return nil
 }
